performance improvement - cache package names
Behavior>>rowanPackageName is sent when building a RowanClassService. #rowanPackageName sends `Rowan image packageNames` as part of it's logic to return the package name for a class.

However, #packageNames  is kind of slow. By caching the #packageNames when building many classes services in one round trip, we gain several orders of magnitude of performance improvement.