get the package/project services from the method service
During a method move involving a hybrid category, instead of trying to weird string parsing to get the package name just get the package & project services from the method service.

We do the update to ensure the state of the package & project dirty flags are up to date in both the "from" and "to" packages.