Check validitiy of process oop before updating processes
The client debugger makes a process service from ready, suspended, and waiting processes while making a process service for it's own process. When a debugger is being killed though, the initial process oop for the debugged process may have been recycled. If the initial process oop is invalid, return an empty collection of processes and let the debugger service die gracefully elsewhere instead of giving a walkback.