save GsProcess during window handle save
Jadeite dumps the oops of its services on the floor each time it makes a round trip. This prevents temp obj space bloat. However, sometimes, it needs to hang onto objects on the server so they aren't garbage collected. It does this by registering the client window handle and the root server object.

During a debugging session, Jadeite needs to hang onto the GsProcess throughout the lifetime of the process or debugging session. Since debugging windows are opened and closed, we don't want to let go of that GsProcess. When a breakpoint is hit, we don't get an error notification so we can save the process for debugging. So if a debugger is opened, and the root object of the window is a GsProcess, do a more persistent save.