Fixed #18 - setting break fails + remove super/sub calc + other  fixes
Setting a breakpoint was not opening a debugger.
When the breakpoint was hit, it was sent to the gci
because RowanMethodService>>setBreakAt: was setting the breakpoint to level 0.
It needed to be set at level 1. Clearing breakpoint had the same problem.

https://github.com/GemTalk/JadeiteForPharo/issues/18

Also tweaked when the active process is suspended when opening a breakpoint. Needed to be done outside the forked block.
Disabled subs and supers implementors calculation in method list. While not used in JfP, it is used in JfD but not relied on. It is suspect in some
out of memory conditions in JfP.
Removed old transcript write which was slowing down ad-hoc workspace executions.